api/requirements.txt
rpi_ws281x
adafruit-circuitpython-neopixel
slackclient
python-dotenv
imutils
rtsp

fastapi
uvicorn[standard]